+++ resources/views/auth/register.blade.php
@@ -75,24 +75,6 @@
 {{-- Inline Js --}}
 @section('inline_js')
 <script>
-
-    $(document).ready(function(){
-        $.notify({
-            icon: "done",
-            title: "<strong>Success</strong>",
-            message: "Coba"
-        }, {
-            type: "success",
-            timer: 150000,
-            delay: 500,
-            newest_on_top: true,
-            placement: {
-                from: "top",
-                align: "right"
-            }
-        });
-    });
-
     $("#registerForm").submit(function(e){ //Prevent default Action for Form
         e.preventDefault();
         formAction();